Update target framework and add package details
Updated the target framework to support both net6.0 and net7.0 in the Menso.Tools.Exceptions project. This will allow the library to be used with a wider range of .NET projects. Additionally, added package details such as PackageId, Authors, Company, PackageDescription, and RepositoryUrl. These changes will improve the package’s ease of identification and traceability on NuGet or other package repositories.