Change Nuget package working directory in deployment workflow
The working directory for packing, publishing, and pushing Nuget packages in the deployment workflow has been updated. Previously, these actions were performed in the ./src/ directory. They are now executed in the root directory (./). This change was necessitated to align with the changes in the project's structure and to ensure the correct path for the Nuget package operations.