Refactor and simplify 'When' class, add new assertion methods
In this commit, the 'When' class is simplified by dividing its functionality among several 'partial' classes based on the type of argument being checked. This makes the code cleaner and easier to manage. Along with that a global using statement for System.Runtime.CompilerServices was added to remove redundant usings from each file separately. The 'IWhen' interface was also updated to include two new methods - AreEqual and AreNotEqual for more granular assertion control. Corresponding changes and tests were also made in 'WhenTests.cs'.