Add [NotNull] attribute to exception methods
The [NotNull] attribute was added to the argument inputs of the exception methods in the IWhen interface, When_GuidType, When_EnumerableType, and When_StringType classes, forcing callers to provide a non-null argument. This change highlights the requirement for a non-null argument more explicitly and helps the prevention of null reference errors. Removed redundant spaces in Menso.Tools.Exceptions.csproj.