Add Scambler library and corresponding unit tests
Scrambler library has been added which includes utilities to shuffle lists of generic types using the Fisher–Yates algorithm. The library includes methods for Shuffle with and without seed, and CryptoStrongShuffle. Accompanying unit tests have also been added to Scrambler.Tests project to validate the correct functioning of these methods. The Menso.Tools.Scrambler was added to the solution and a readme file was created to document its use.