🧹 (Exceptions) Split IWhen interface into multiple specific interfaces | closes #15
The IWhen interface was getting quite large and was dealing with a wide variety of cases, including Strings, Guids, Booleans, and Enumerables. To make it more manageable and follow the Single Responsibility Principle, we have decided to split it into multiple specific interfaces. It will make the code more readable, maintainable and easier to work with in the future. Now, the IWhen interface simply integrates all these smaller interfaces.