🐞 (Exceptions) Renamed method 'FalseOrNull' to 'NullOrFalse' | closes #17
This change was made to improve the readability and clarity of the code. The method name 'NullOrFalse' is more intuitive and aligns better with its actual functionality, as it checks if the argument is null or false. This way, other developers can understand the code more easily without having to dig into its implementation.