Add Github Action for running tests
This commit introduces a new Github Action that automatically runs tests for each pull request made to the 'main' branch. Utilizing the 'ubuntu-latest' runner, the workflow first builds and test the code with 'dotnet' with the results stored in the '/TestResults' directory. It then parses the test results using the 'trx-parser' action. This ensures all tests are run and results are parsed for every change, enhancing code reliability.