Update packages and add CI flag
The changes in this commit address two primary goals:

Firstly, the version numbers of the named packages Microsoft.NET.Test.Sdk, xunit, and xunit.runner.visualstudio have been increased. The upgrade offers enhanced features, bug fixes and potential performance improvements that can benefit the project.

Secondly, a ContinuousIntegrationBuild property with the condition "'$(GITHUB_ACTIONS)' == 'true'" was added to the csproj files. This was done to make sure that the build process behaves appropriately when it's running on a Continuous Integration server like GitHub Actions.