"Refactor test and coverage workflows into separate files"
Separated test and coverage operations into two different workflow files for the sake of simplicity and to assign different triggers for each workflow. This way, we can run coverage only on 'main' branch after push events and not every time we run the tests. This will reduce unnecessary action runs and provide a clearer structure to our CI/CD pipeline.